Attacking , damage dealing & MessageArgs
"Health" is added to everything that can get damaged (except from the player).
On biting animations, attacking colliers send an attack message to every wanted object that gets triggerred.
If attacking message gets recieved from a "Health" script, that it calls "DealDamageTo" with itself as parameter. This send message systm allows the attack to not mean only damage but for other uses too, for example to bite objects and move them.
"MessageArgs" was added. SendMessage will send this object with values: sender, recieved.