Cake tongue movement and animation
Tongue animations left and right was changed to add more range.
For the tongue aiming: Mouse potition is converted from screen to world potition with offset relative to the cake. This vector3 is then converted to local position difference. Then this values x,y are inputs to the bland tree x,y with range of x(-5,5) , y(-2,2). This whole method allows to aim with the tongue with all cameras and not rely on screen viewport.