Tongue attack collier and animations
Acollider was added with "AttackCollider.cs" that damages specifie objects upon trigger.
An enum "AttackType" was added on Attack.cs that is used to define the attack type that was triggered to deal the proper damage (Bite, TongueSwing).
AttackMessageArgs.cs was added to deliver info about the attacks.
All 5 of tongue swinging animations are shorter to play quicker and don't give a delay feal on attack.