Attack colliders, drag and rolling.
Cake's rigidbody air drag was increased to 1.
Attack colliders on bite and tongue swing are disabled and are only enabled on the attack animations.
Roll adjustments are faster on locked camera.
On locked camera and when the character is facing the camera and rolls, the cake rolls to the other direction. This doesn't change the final diraction of the roll. If the players presses to roll left it to his right but on the cameras left.